+++ x/ref/runtime/internal/rpc/server.go
@@ -27,12 +27,12 @@ import (
 	"v.io/v23/verror"
 	"v.io/v23/vom"
 	"v.io/v23/vtrace"
+	"v.io/v23/requestid"
 	"v.io/x/ref/lib/publisher"
 	"v.io/x/ref/lib/pubsub"
 	"v.io/x/ref/lib/stats"
 	"v.io/x/ref/runtime/internal/flow/conn"
 	"v.io/x/ref/runtime/internal/flow/manager"
-	"v.io/x/ref/runtime/internal/requestid"
 )
 
 const (
